requirements.txt
django>=4.2
djangorestframework>=3.14
torch
torchvision
detectron2 -f https://dl.fbaipublicfiles.com/detectron2/wheels/cu118/torch2.0/index.html
pillow

